+++ src/store/features/notification/notificationService.tsx
@@ -1,14 +1,14 @@
-/* eslint-disable */
-import {axiosInstance} from "../../../utils/axios/axiosInstance";
+// /* eslint-disable */
+// import {axiosInstance} from "../../../utils/axios/axiosInstance";
 
-const fetchNotifications = async () => {
+// const fetchNotifications = async () => {
 
-    const response = await axiosInstance.get(`/api/user/user-get-notifications`);
-    return response.data;
-};
+//     const response = await axiosInstance.get(`/api/user/user-get-notifications`);
+//     return response.data;
+// };
 
-const notificationService = {
-  fetchNotifications,
-};
+// const notificationService = {
+//   fetchNotifications,
+// };
 
-export default notificationService;
+// export default notificationService;
